+++ src/pybx/anchor.py
@@ -5,6 +5,7 @@ import numpy as np
 from fastcore.foundation import L
 
 from .ops import __ops__, get_op, named_idx
+from .basics import get_bx
 
 
 def get_edges(image_sz: tuple, feature_sz: tuple, op='noop'):
@@ -30,6 +31,26 @@ def get_edges(image_sz: tuple, feature_sz: tuple, op='noop'):
     mesh = np.meshgrid(x_, y_)
     edges = np.stack([m.flatten() for m in mesh], axis=-1)
     return edges
+
+
+def validate_boxes(coords, image_sz, feature_sz, labels=None, clip=True, min_vis=0.25):
+    """Validate calculated anchor box coords.
+    :param coords: anchor box coordinates
+    :param labels: anchor box labels
+    :param image_sz: tuple of (width, height) of an image
+    :param feature_sz: tuple of (width, height) of a channel
+    :param clip: whether to apply np.clip
+    :param min_vis: minimum visibility dictates the condition for a box to be considered valid. The value corresponds to the
+    ratio of expected area to the calculated area after clipping to image dimensions.
+    :return: anchor box coordinates in [pascal_voc] format
+    """
+    _max = max(image_sz[0], image_sz[1])
+    # clip the boxes to image dimensions
+    b = get_bx(coords.clip(0, _max), labels) if clip else get_bx(coords, labels)
+    # check if the area of the bounding box is fitting the minimum area criterion
+    min_area = (image_sz[0] / feature_sz[0]) * (image_sz[1] / feature_sz[1]) * min_vis
+    b = get_bx([b_.values() for b_ in b if b_.area() > min_area])
+    return b
 
 
 def bx(image_sz: tuple, feature_sz: tuple, asp_ratio: float = None, clip=True, named=True, anchor_sfx: str = 'a', min_vis=0.25):
